backend_azeri_student/requirements.txt
amqp==2.6.1
arabic-reshaper==2.1.3
arrow==1.2.1
asgiref==3.5.0
aspy.refactor-imports==2.2.0
autoflake==1.4
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.2
billiard==3.6.4.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.0
chardet==3.0.4
click==8.0.3
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.12
django-3-jet==1.0.8
django-admin-rangefilter==0.6.2
django-ajax-selects==1.9.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-debug-toolbar==2.2
django-excel-response
django-filter==21.1
django-import-export==2.7.1
django-js-asset==1.2.2
django-map-widgets==0.3.3
django-mathfilters==1.0.0
django-modeltranslation==0.15.1
django-multiselectfield==0.1.12
django-nested-inline==0.4.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rosetta==0.9.4
django-social-auth==0.7.28
django-timezone-field==4.1.2
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docopt==0.6.2
drf-yasg==1.17.1
et-xmlfile==1.1.0
facebook-sdk==3.1.0
Faker==4.14.2
feedparser==6.0.8
filelock==3.0.12
firebase-admin==4.5.0
flake8==3.9.2
future==0.18.2
gitlint==0.17.0
gitlint-core==0.17.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
identify==2.2.12
idna==2.10
importlib-metadata==4.10.1
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
kombu==4.6.11
libretranslatepy==2.1.1
lxml==4.6.3
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.20.2
oauth2==1.9.0.post1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==20.9
pandas==1.1.2
pdfkit==0.6.1
Pillow==7.2.0
platformdirs==2.2.0
polib==1.1.1
pre-commit==2.13.0
proto-plus==1.18.1
protobuf==3.15.8
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycountry==20.7.3
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.3.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-barcode==0.13.1
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
raven==6.10.0
redis==3.5.3
reorder-python-imports==2.6.0
reportlab==3.6.6
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sendgrid==6.4.8
sgmllib3k==1.0.0
sh==1.14.2
simple-crypt==4.1.7
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.2.1
sqlparse==0.4.2
starkbank-ecdsa==1.1.0
tablib==3.2.0
terminaltables==3.1.10
text-unidecode==1.3
toml==0.10.2
translate==3.6.1
typing-extensions==3.10.0.0
Unidecode==1.1.0
uritemplate==3.0.1
urllib3==1.25.11
uWSGI==2.0.19.1
vine==1.3.0
virtualenv==20.7.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.6
xlwt==1.3.0
zipp==3.7.0
